Renaming :collection to :collection_map
When I was looking at the Lotus::Model::Adapters::Sql::Collection, I
was confused about collection variable name. Given that there is a
collection of maps and an adapter collection, I found changing the
variable name to be easer to understand.

As this is buried in the internals of implementation, I believe it is
fair to change.

This is also a potential conversation, as in naming it a
collection_map I am removing some abstraction in name towards a
concretion.