Add jRuby 9k support
It introduces a DB URI connection wrapper where we can hold logic of how to
deal with DB connection URLs independent from platforms.

- [x] Add jdbc libraries for jruby platform
- [x] Extract DB URI connection logic from Adapter
- [x] Add support for the current SQL adapters
  - [x] SQLite
  - [x] PostgreSQL
  - [x] MySQL
- [x] Improve `execute` interface do mimic Sequel adapter

----------------------------------------------------------------------

Important
---------

When using `execute` directly we're adrift of the adapter, when it is
the Ruby SQlite3 adapter we get `SQlite3::ResultSet` as result,
but when dealing with Java JDBC SQLite adapter with get
`Java::OrgSqliteJdbc4::JDBC4ResultSet` which is a low level object that
doesn't have the Ruby interface that we expect.

Facing that we need to question ourselves if we want to expose the raw
adapter reponse or a object with a cross-platform normalized interface.

Remove outdated jRuby 9000 on travis, use head

Add spare time to avoid race condition

Enable install on .travis.yml

Try to refresh Travis cache

Install deps :-P

WHY U DO NOT UPDATE RVM?!