Adds simple BCrypt secure passwords
Includes basic validation for password and confirmation. Ensures
passwords are kept below 72 characters, the theoretical maximum
permitted by bcrypt's hashing function.

I hit some troubles with Lotus validations. It was a surprise for me to
find that the validator macro creates accessors/attributes. It would
have been tidier to include a module containing the instance methods
within the `included` callback, but due to ordering with the attributes
API, and my need to overload the `password=` accessor _after_ it had
been defined by the attributes API.

Because I wanted to use the confirmation, and the `size` validators
this pushed me toward that particular approach.

Some things to note:

* BCrypt is latently loaded and needs to be present in the Gemfile,
 this will be documented of course.
* No efforts were made to abstract access to BCrypt since it's just the
 defacto approach/library. Happy to do this though if anyone feels
 strongly.
* At the moment, the hashing cost is whatever BCrypt's default is. This
 can be overridden via `BCrypt::Engine.cost=`. It may be worth
 providing a nicer API here for consumers.

I just wanted to get some validation on the approach before I
documented the code. WDYT?