Timestamp plugin respects given timestamps.
If no timestamps is given, as it most often will, then current time
will be used. If timestamps are given, the given timestamps will be
used.

This restores pre-0.7 behaviour and I think it is correct that when you
ask the repository to insert/update a row with given data it shouldn’t
change given data before data is persisted to the database.