Reliably parse query params from connection string (#485)
The previous regex used the match query options was incorrect since it only
captured words (`\w+`) values from query params. This would cause params with
values such as `&host=/path/to/my/socket` to be parsed as `nil`, which is
incorrect and will cause pain on some hosting providers (i.e. Google Cloud
requires a path to a socket).

Parsing query params can be brittle, so we opt to let `CGI.parse` to do the
heavy lifting since it is more battle-tested.