Add #one to HasMany (#550)
Per https://github.com/hanami/model/issues/550, the HasMany interface
does not line up with Hanami's guides. Prior to this commit, the
following is invalid as HasMany does not implement #one.

```ruby
class AuthorRepository < Hanami::Repository
  associations do
    has_many :books
  end

  def find_book(author, id)
    book_for(author, id).one
  end

  private

  def book_for(author, id)
    assoc(:books, author).where(id: id)
  end
end
```

This implementation is based on a couple of others, notably ROM's and
Ecto's (Elixir). I opted to name the raised error MultipleResultsError,
which is the exact name used by Ecto, because I feel it bears a semantic
relationship to the name of the method from which raises it.