Target Updated Default Travis CI Build Environment (#556)
* Update Postgres version

Certain (Postgres-specific) specs are expecting syntax specific to
versions newer versions of Postgres (10 and up). TravisCI has been
running 9.4, which does not support this syntax.

This commit is specifically targeting the following failure, where `as
integer` is valid syntax in versions 10 and higher.

```bash
  1) Hanami::Model::Migrator PostgreSQL apply dumps database schema.sql
     Failure/Error:
       expect(actual).to include <<~SQL
         CREATE SEQUENCE public.reviews_id_seq
             AS integer
             START WITH 1
             INCREMENT BY 1
             NO MINVALUE
             NO MAXVALUE
             CACHE 1;
       SQL
```

* Start MySQL server on build

As of 4/2019, Travis moved to a new distribution of Ubuntu (Xenial). Per
their announcement*, "Services like MySQL or PostgreSQL are not started
by default."

This lines up with the error message in CI:

```
ERROR 2002 (HY000): Can't connect to local MySQL server through socket '/var/run/mysqld/mysqld.sock' (2)
```

Adding the MySQL service to resolve this.

https://blog.travis-ci.com/2019-04-15-xenial-default-build-environment#1-services-support