Use environment variables when interacting with PostgreSQL CLI tools (#622)
Instead of using command line flags, environment variables are set for the execution of
any PostgreSQL command.

The migrator spec builds a URL that excludes the password, but that URL is then used to
build a connection, which is then used to run a PostgreSQL command. That command will
not receive a password because the password in the connection string was absent.

Remove the ci.yml hack that forces PGPASSWORD to be present.